+++ lib/element/src/main/java/com/nl/lib/element/skill/SkillEntity.kt
@@ -19,6 +19,6 @@ data class SkillEntity(
     val desc: String,                             // 技能描述
     val level: Int = 1,                           // 技能等级
     val maxLevel: Int = 10,                       // 技能最大等级
-    val triggerProbability: Float = 0.3F,           // 技能触发概率
-    val triggerProbabilityGrowth: Float = 0.03F,   // 技能触发概率成长值
+    val triggerProbability: Double = 0.3,           // 技能触发概率
+    val triggerProbabilityGrowth: Double = 0.03,   // 技能触发概率成长值
 )
